Migrate EventHelper to TS
